% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_code.R
\name{add_snippet}
\alias{add_snippet}
\title{Convert code to code snippet}
\usage{
add_snippet(
  script = system.file("conversion_staging.R", package = "advancedShiny")
)
}
\arguments{
\item{script}{the R script to convert to a code snippet}
}
\value{
inserts the code necessary as a code snippet below the current selection
}
\description{
Convert code to code snippet
}
