% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{code_block}
\alias{code_block}
\title{Render code i.e. package (or function) name}
\usage{
code_block(string, pkg = FALSE, href = NULL, noWS = NULL)
}
\arguments{
\item{pkg}{logical, if TRUE the output text will be italicised (representing a package name). If FALSE it
won't be italicised (representing a function name or arbitrary R code).}

\item{href}{optional link to include}

\item{code}{the text to be formatted}
}
\value{
a HTML string which can be inserted directly into the UI of the app, showing the
code as formatted, (potentially) clickable text
}
\description{
Renders either a package name (as italics and formatted as code) or a function (formatted as code),
optionally include a link
}
