% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\name{get_shinydashboard_theme}
\alias{get_shinydashboard_theme}
\title{Custom shinydashboard theme}
\usage{
get_shinydashboard_theme()
}
\value{
a shinydashboard theme created using fresh::create_theme
}
\description{
Customise some visual elements of the dashboard using the *fresh* package.
This function returns a theme which is passed to use_theme, in the dashboardBody
element of the UI.
}
