% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_treant.R
\name{node}
\alias{node}
\title{Define treant tree node}
\usage{
node(name, desc, children = NULL, ...)
}
\arguments{
\item{name}{the name of the node}

\item{desc}{the description of the node}

\item{children}{a list of children nodes}
}
\description{
Helper function for defining tree structure, helps
to create trees/nodes with less code
}
