% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_IAS_app}
\alias{run_IAS_app}
\title{Run the app}
\usage{
run_IAS_app()
}
\value{
either run the app as a side effect or return a shiny.appobj object
}
\description{
Initiate an instance of the 'Intermediate & Advanced Shiny' app. If the current context is interactive,
(i.e. if the function is called from RStudio), this function will initiate the app stored in
inst/app with the runApp function. If the context is not interactive (i.e. the project is being sourced in RStudio
Connect after having been deployed), then the function will return a shiny.appobj object, using the function
shinyAppDir. This distinction is necessary to allow the app to be deployed.
}
\details{
Note: an alias function 'l' is also provided to speed up workflow (after making changes, load them with 'Ctrl+Shift+L',
then launch the app with 'l()')
}
