% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_15_appStructure.R
\name{tab_appStructure_server}
\alias{tab_appStructure_server}
\title{App Structure module server}
\usage{
tab_appStructure_server(id, appData)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the App Structure module
}
