% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_15_appStructure.R
\name{tab_appStructure_ui}
\alias{tab_appStructure_ui}
\title{App Structure module
App Structure module UI}
\usage{
tab_appStructure_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the App Structure module
}
