% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_23_attaliverse.R
\name{tab_attaliverse_ui}
\alias{tab_attaliverse_ui}
\title{Attali-verse module
Attali-verse module UI}
\usage{
tab_attaliverse_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Attali-verse module
}
