% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_13_controllingReactivity.R
\name{tab_controllingReactivity_server}
\alias{tab_controllingReactivity_server}
\title{Controlling Reactivity module server}
\usage{
tab_controllingReactivity_server(appData)
}
\arguments{
\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Controlling Reactivity module
}
