% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_14_controllingReactivity.R
\name{tab_controllingReactivity_ui}
\alias{tab_controllingReactivity_ui}
\title{Controlling Reactivity module
Controlling Reactivity module UI}
\usage{
tab_controllingReactivity_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Controlling Reactivity module
}
