% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_25_customInputs.R
\name{tab_customInputs_server}
\alias{tab_customInputs_server}
\title{Custom Shiny Inputs module server}
\usage{
tab_customInputs_server(id, appData)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Custom Shiny Inputs module
}
