% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_25_customInputs.R
\name{tab_customInputs_ui}
\alias{tab_customInputs_ui}
\title{Custom Shiny Inputs module
Custom Shiny Inputs module UI}
\usage{
tab_customInputs_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Custom Shiny Inputs module
}
