% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_35_errorHandling.R
\name{tab_errorHandling_ui}
\alias{tab_errorHandling_ui}
\title{Error Handling & Debugging module
Error Handling & Debugging module UI}
\usage{
tab_errorHandling_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Error Handling & Debugging module
}
