% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_24_extendingShiny.R
\name{tab_extendingShiny_server}
\alias{tab_extendingShiny_server}
\title{Extending Shiny module server}
\usage{
tab_extendingShiny_server(id, appData)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Extending Shiny module
}
