% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_24_extendingShiny.R
\name{tab_extendingShiny_ui}
\alias{tab_extendingShiny_ui}
\title{Extending Shiny module
Extending Shiny module UI}
\usage{
tab_extendingShiny_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Extending Shiny module
}
