% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_00_info.R
\name{tab_info_ui}
\alias{tab_info_ui}
\title{Info module
Info module UI}
\usage{
tab_info_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Info module
}
