% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_x_template.R
\name{tab_modCCx_ui}
\alias{tab_modCCx_ui}
\title{modTCx module
modTCx module UI}
\usage{
tab_modCCx_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the modTCx module
}
