% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_36_optimisation.R
\name{tab_optimisation_server}
\alias{tab_optimisation_server}
\title{Optimisation module server}
\usage{
tab_optimisation_server(appData)
}
\arguments{
\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Optimisation module
}
