% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_36_optimisation.R
\name{tab_optimisation_ui}
\alias{tab_optimisation_ui}
\title{Optimisation module
Optimisation module UI}
\usage{
tab_optimisation_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Optimisation module
}
