% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_12_reactiveGraph.R
\name{tab_reactiveGraph_ui}
\alias{tab_reactiveGraph_ui}
\title{Reactive Graph module
Reactive Graph module UI}
\usage{
tab_reactiveGraph_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Reactive Graph module
}
