% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_14_reactiveObjects.R
\name{tab_reactiveObjects_server}
\alias{tab_reactiveObjects_server}
\title{Reactive Objects module server}
\usage{
tab_reactiveObjects_server(id, appData)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Reactive Objects module
}
