% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_14_reactiveObjects.R
\name{tab_reactiveObjects_ui}
\alias{tab_reactiveObjects_ui}
\title{Reactive Objects module
Reactive Objects module UI}
\usage{
tab_reactiveObjects_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Reactive Objects module
}
