% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_40_references.R
\name{tab_references_server}
\alias{tab_references_server}
\title{References & Further Materials module server}
\usage{
tab_references_server(id, appData)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the References & Further Materials module
}
