% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_4_references.R
\name{tab_references_ui}
\alias{tab_references_ui}
\title{References & Further Materials module
References & Further Materials module UI}
\usage{
tab_references_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the References & Further Materials module
}
