% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_31_sessionObject.R
\name{tab_sessionObject_ui}
\alias{tab_sessionObject_ui}
\title{Session Object module
Session Object module UI}
\usage{
tab_sessionObject_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Session Object module
}
