% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_16_shinyModules.R
\name{tab_shinyModules_server}
\alias{tab_shinyModules_server}
\title{Shiny Modules module server}
\usage{
tab_shinyModules_server(id, appData)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Shiny Modules module
}
