% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_16_shinyModules.R
\name{tab_shinyModules_ui}
\alias{tab_shinyModules_ui}
\title{Shiny Modules module
Shiny Modules module UI}
\usage{
tab_shinyModules_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Shiny Modules module
}
