% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_32_trackingUsage.R
\name{tab_trackingUsage_ui}
\alias{tab_trackingUsage_ui}
\title{Tracking Usage module
Tracking Usage module UI}
\usage{
tab_trackingUsage_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Tracking Usage module
}
