% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_12_understandingReactivity.R
\name{tab_understandingReactivity_ui}
\alias{tab_understandingReactivity_ui}
\title{Understanding Reactivity module
Understanding Reactivity module UI}
\usage{
tab_understandingReactivity_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Understanding Reactivity module
}
