% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_21_usingCSS.R
\name{tab_usingCSS_server}
\alias{tab_usingCSS_server}
\title{Using CSS module server}
\usage{
tab_usingCSS_server(appData)
}
\arguments{
\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Using CSS module
}
