% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_21_usingCSS.R
\name{tab_usingCSS_ui}
\alias{tab_usingCSS_ui}
\title{Using CSS module
Using CSS module UI}
\usage{
tab_usingCSS_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Using CSS module
}
