% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_22_usingJS.R
\name{tab_usingJS_server}
\alias{tab_usingJS_server}
\title{Using Javascript module server}
\usage{
tab_usingJS_server(id, appData)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Using Javascript module
}
