% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_22_usingJS.R
\name{tab_usingJS_ui}
\alias{tab_usingJS_ui}
\title{Using Javascript module
Using Javascript module UI}
\usage{
tab_usingJS_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Using Javascript module
}
