% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_13_usingReactivity.R
\name{tab_usingReactivity_ui}
\alias{tab_usingReactivity_ui}
\title{Using Reactivity module
Using Reactivity module UI}
\usage{
tab_usingReactivity_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Using Reactivity module
}
