% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_33_visualisationPlotly.R
\name{tab_visualisationPlotly_server}
\alias{tab_visualisationPlotly_server}
\title{Visualisation with plotly module server}
\usage{
tab_visualisationPlotly_server(appData)
}
\arguments{
\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Visualisation with plotly module
}
