% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_33_visualisationPlotly.R
\name{tab_visualisationPlotly_ui}
\alias{tab_visualisationPlotly_ui}
\title{Visualisation with plotly module
Visualisation with plotly module UI}
\usage{
tab_visualisationPlotly_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Visualisation with plotly module
}
