% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_34_visualisationR2D3.R
\name{tab_visualisationR2D3_server}
\alias{tab_visualisationR2D3_server}
\title{Visualisation with r2d3 module server}
\usage{
tab_visualisationR2D3_server(id, appData)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Visualisation with r2d3 module
}
