% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_34_visualisationR2D3.R
\name{tab_visualisationR2D3_ui}
\alias{tab_visualisationR2D3_ui}
\title{Visualisation with r2d3 module
Visualisation with r2d3 module UI}
\usage{
tab_visualisationR2D3_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Visualisation with r2d3 module
}
