% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_11_whatIsShiny.R
\name{tab_whatIsShiny_server}
\alias{tab_whatIsShiny_server}
\title{What Is Shiny? module server}
\usage{
tab_whatIsShiny_server(appData)
}
\arguments{
\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the What Is Shiny? module
}
