% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_11_whatIsShiny.R
\name{tab_whatIsShiny_ui}
\alias{tab_whatIsShiny_ui}
\title{What Is Shiny? module
What Is Shiny? module UI}
\usage{
tab_whatIsShiny_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the What Is Shiny? module
}
