% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_treant.R
\name{treant}
\alias{treant}
\title{treant htmlwidget}
\usage{
treant(structure, width = NULL, height = NULL)
}
\arguments{
\item{structure}{a list of lists denoting the structure of the tree to be built}
}
\description{
interface to the Treant javascript package
}
