% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_35_errorHandling.R
\name{tab_errorHandling_server}
\alias{tab_errorHandling_server}
\title{Error Handling & Debugging module server}
\usage{
tab_errorHandling_server(appData)
}
\arguments{
\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Error Handling & Debugging module
}
