% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_12_reactiveGraph.R
\name{tab_reactiveGraph_server}
\alias{tab_reactiveGraph_server}
\title{Reactive Graph module server}
\usage{
tab_reactiveGraph_server(appData)
}
\arguments{
\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Reactive Graph module
}
