% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_31_sessionObject.R
\name{tab_sessionObject_server}
\alias{tab_sessionObject_server}
\title{Session Object module server}
\usage{
tab_sessionObject_server(appData)
}
\arguments{
\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Session Object module
}
